+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [DefaultWith](https://fsharp.github.io/fsharp-core-docs/reference/fsharp-core-optionmodule.html#defaultWith) function
 - [Exists](https://fsharp.github.io/fsharp-core-docs/reference/fsharp-core-optionmodule.html#exists) function
 - [Filter](https://fsharp.github.io/fsharp-core-docs/reference/fsharp-core-optionmodule.html#filter) function
+- [Flatten](https://fsharp.github.io/fsharp-core-docs/reference/fsharp-core-optionmodule.html#flatten) function
 ### Changed
 - Class structure. Moved Ext class to own file
 - Test project target framework to .Net 6